커버링 인덱스 방식으로 페이지 구현

+++ src/main/java/com/kakao/cafe/domain/PageSelector.java
@@ -14,7 +14,7 @@ public class PageSelector {
     private final Page nextPage;
     private final boolean hasNext;
 
-    public PageSelector(final int page, final int totalArticles) {
+    public PageSelector(final int page, final long totalArticles) {
         Checker.checkPage(page);
         final int maxPage = (int)Math.ceil((double)totalArticles / Page.MAX_ARTICLES);
         Checker.checkPage(maxPage);
